+++ tests/Examples/search.py
